prod-etl-pipeline-main/requirements.txt
# For data generation
faker
pandas
pyspark

# For GCP interaction
google-cloud-storage
google-cloud-dataproc

# For testing
pytest
pylint